modefy demo.js

+++ demo.js
@@ -3,13 +3,19 @@ var fs = require('fs')
 var dirName = process.argv[2] 
 
 process.chdir("/Users/Sam/Desktop") 
-fs.mkdirSync("./" + dirName) 
-process.chdir("./" + dirName) 
-fs.mkdirSync('css') 
-fs.mkdirSync('js') 
 
-fs.writeFileSync("./index.html", "\<!DOCTYPE\>\n\<title\>Hello\<\/title\>\n\<h1\>Hi\<\/h1\>")
-fs.writeFileSync("css/style.css", "h1{color: red;}")
-fs.writeFileSync("./js/main.js", 'var string = "Hello World"\nalert(string)')
+if (fs.existsSync("./" + dirName)) {
+    console.log("dir was exist")
+    process.exit(1)
+} else {
+    fs.mkdirSync("./" + dirName) 
+    process.chdir("./" + dirName) 
+    fs.mkdirSync('css') 
+    fs.mkdirSync('js') 
 
-process.exit(0)
+    fs.writeFileSync("./index.html", "\<!DOCTYPE\>\n\<title\>Hello\<\/title\>\n\<h1\>Hi\<\/h1\>")
+    fs.writeFileSync("css/style.css", "h1{color: red;}")
+    fs.writeFileSync("./js/main.js", 'var string = "Hello World"\nalert(string)')
+
+    process.exit(0)
+}
